requirements.txt
beautifulsoup4==4.12.2
fastapi==0.98.0
itsdangerous==2.1.2
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
slowapi==0.1.8
uvicorn==0.22.0
